More work for `icon`
Introduce a function that converts an NSImage to a PNG and returns its path to a temporary location on disk.

This way, NPAuthenticator can load the icon, and the user can provide a cool icon himself.  (Later, he won't have to... NSAuthenticatedTask will be able to find itself the correct icon.)